Implementado Arquivo Conftest
close #44
