+++ libpythonprojat2022/spam/db.py
@@ -22,7 +22,7 @@ class Sessao:
 
 class Conexao:
     def __init__(self):
-        sleep(10)
+        sleep(1)
 
     def gerar_sessao(self):
         return Sessao()
