+++ libpythonprojat2022/tests/test_spam/test_envio_para_base_de_usuario.py
@@ -1,0 +1,11 @@
+from libpythonprojat2022.spam.enviador_de_email import Enviador
+from libpythonprojat2022.spam.main import EnviadorDeSpam
+
+
+def test_envio_de_spam(sessao):
+    enviador_de_spam = EnviadorDeSpam(sessao, Enviador())
+    enviador_de_spam.enviar_emails(
+        'renzo@python.pro.br',
+        'Curso Python Pro',
+        'Confira os modulos fantasticos'
+    )
